api_getwaye/requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
djangorestframework==3.16.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.
